+++ src/view/taskView.ts
@@ -8,6 +8,7 @@ import { resignFn } from '../utils/utils';
 function getTaskInfoItems(task: Task): vscode.TreeItem[] {
     return [
         new InfoItem(task.jobid.toString(), 'id'),
+        new InfoItem(task.node, 'nodelist'),
         new InfoItem(task.gres?.toString() ?? 'No GRES', 'GRES'),
         new InfoItem(task.command, 'command'),
         new LogFileItem(task.out_path, 'stdout'),